responses: uppercase letter, messages and special character because database latin encoding. models: comments. routes: response codes, CRUD and PATCH structures.

+++ api/routes/customers.py
@@ -25,8 +25,10 @@ def allcustomers():
 
 @customer_routes.route('/customers/', methods=['POST'])
 def postcustomer():
+    data = request.get_json()
+    if not data:
+        return resp.response_with(resp.BAD_REQUEST_400), resp.BAD_REQUEST_400
     try:
-        data = request.get_json()
         row = object_schema.load(data)
         session.add(row)
         session.commit()
@@ -41,22 +43,50 @@ def postcustomer():
 def getcustomer(customerNumber):
     found = customerNumber
     row = session.query(Customer).get(found)
-    result = object_schema.dump(row)
-    if result:
-        return resp.response_with(resp.SUCCESS_200, value={"Request": result}), resp.SUCCESS_200        # "......!!!"
-    else:
+    if not row:
         return resp.response_with(resp.SERVER_ERROR_404), resp.SERVER_ERROR_404
+    else:
+        result = object_schema.dump(row)
+        return resp.response_with(resp.SUCCESS_200, value={"Request": result}), resp.SUCCESS_200
 
 
 @customer_routes.route('/customers/<int:customerNumber>', methods=['PUT'])
 def putcustomer(customerNumber):
     found = customerNumber
     data = request.get_json()
-    result = object_schema.dump(data)
+    if not data:
+        return resp.response_with(resp.BAD_REQUEST_400), resp.BAD_REQUEST_400
     try:
-        session.query(Customer).filter(Customer.customerNumber==found).update(result)
+        row = object_schema.dump(data)
+        session.query(Customer).filter(Customer.customerNumber==found).update(row)
         session.commit()
-        return resp.response_with(resp.SUCCESS_200, value={"Updated Row": result}), resp.SUCCESS_200
+        result = object_schema.dump(session.query(Customer).get(found))
+        return resp.response_with(resp.SUCCESS_201, value={"Updated Row": result}), resp.SUCCESS_201
+    except IntegrityError as error:
+        session.rollback()
+        return resp.response_with(resp.BAD_REQUEST_400), resp.BAD_REQUEST_400
+
+
+@customer_routes.route('/customers/<int:customerNumber>', methods=['PATCH'])
+def patchoffice(customerNumber):
+    found = customerNumber
+    data = request.get_json()
+    if not data:
+        return resp.response_with(resp.BAD_REQUEST_400), resp.BAD_REQUEST_400
+    try:
+        row = session.query(Customer).get(found)
+        if data.get('phone'):
+            row.phone = data['phone']
+        if data.get('addressLine1'):
+            row.addressLine1 = data['addressLine1']
+        if data.get('addressLine2'):
+            row.addressLine2 = data['addressLine2']
+        if data.get('postalCode'):
+            row.postalCode = data['postalCode']
+        session.add(row)
+        session.commit()
+        result = object_schema.dump(row)
+        return resp.response_with(resp.SUCCESS_200, value={"Updated Row Fields": result}), resp.SUCCESS_200
     except IntegrityError as error:
         session.rollback()
         return resp.response_with(resp.BAD_REQUEST_400), resp.BAD_REQUEST_400
@@ -65,11 +95,13 @@ def putcustomer(customerNumber):
 @customer_routes.route('/customers/<int:customerNumber>', methods=['DELETE'])
 def delcustomer(customerNumber):
     found = customerNumber
+    row = session.query(Customer).get(found)
+    if not row:
+        return resp.response_with(resp.SERVER_ERROR_404), resp.SERVER_ERROR_404
     try:
-        row = session.query(Customer).get(found)
         session.delete(row)
         session.commit()
-        return resp.response_with(resp.SERVER_ERROR_404), resp.SERVER_ERROR_404
+        return resp.response_with(resp.SUCCESS_204, value={'message': 'Deleted Row'}), resp.SUCCESS_204
     except SQLAlchemyError as error:
         session.rollback()
         return resp.response_with(resp.BAD_REQUEST_400), resp.BAD_REQUEST_400
