+++ api/routes/orderdetails.py
@@ -19,14 +19,16 @@ session = Session()
 @orderdetail_routes.route('/orderdetails/', methods=['GET'])
 def allorderdetails():
     rows = session.query(Orderdetail).all()
-    result = object_schema.dump(rows, True)
+    result = object_schema.dump(rows, many=True)
     return resp.response_with(resp.SUCCESS_200, value={"Row List": result}), resp.SUCCESS_200
 
 
 @orderdetail_routes.route('/orderdetails/', methods=['POST'])
 def postorderdetail():
+    data = request.get_json()           # review datetime
+    if not data:
+        return resp.response_with(resp.BAD_REQUEST_400), resp.BAD_REQUEST_400
     try:
-        data = request.get_json()           # review datetime
         row = object_schema.load(data)
         session.add(row)
         session.commit()
@@ -41,22 +43,45 @@ def postorderdetail():
 def getorderdetails(orderNumber):
     found = orderNumber
     row = session.query(Orderdetail).get(found)
-    result = object_schema.dump(row)
-    if result:
-        return resp.response_with(resp.SUCCESS_200, value={"Request": result}), resp.SUCCESS_200
-    else:
+    if not row:
         return resp.response_with(resp.SERVER_ERROR_404), resp.SERVER_ERROR_404
+    else:
+        result = object_schema.dump(row)
+        return resp.response_with(resp.SUCCESS_200, value={"Request": result}), resp.SUCCESS_200
 
 
 @orderdetail_routes.route('/orderdetails/<int:orderNumber>', methods=['PUT'])
 def putorderdetails(orderNumber):
     found = orderNumber
     data = request.get_json()
-    result = object_schema.dump(data)
+    if not data:
+        return resp.response_with(resp.BAD_REQUEST_400), resp.BAD_REQUEST_400
     try:
-        session.query(Orderdetail).filter(Orderdetail.orderNumber==found).update(result)
+        row = object_schema.dump(data)
+        session.query(Orderdetail).filter(Orderdetail.orderNumber==found).update(row)
         session.commit()
-        return resp.response_with(resp.SUCCESS_200, value={"Updated Row": result}), resp.SUCCESS_200
+        result = object_schema.dump(session.query(Orderdetail).get(found))
+        return resp.response_with(resp.SUCCESS_201, value={"Updated Row": result}), resp.SUCCESS_201
+    except IntegrityError as error:
+        session.rollback()
+        return resp.response_with(resp.BAD_REQUEST_400), resp.BAD_REQUEST_400
+
+
+@orderdetail_routes.route('/orderdetails/<int:orderNumber>', methods=['PATCH'])
+def patchoffice(orderNumber):
+    found = orderNumber
+    data = request.get_json()
+    if not data:
+        return resp.response_with(resp.BAD_REQUEST_400), resp.BAD_REQUEST_400
+    try:
+        row = session.query(Orderdetail).get(found)
+        if data.get('quantityOrdered'):
+            row.quantityOrdered = data['quantityOrdered']
+
+        session.add(row)
+        session.commit()
+        result = object_schema.dump(row)
+        return resp.response_with(resp.SUCCESS_200, value={"Updated Row Fields": result}), resp.SUCCESS_200
     except IntegrityError as error:
         session.rollback()
         return resp.response_with(resp.BAD_REQUEST_400), resp.BAD_REQUEST_400
@@ -65,11 +90,13 @@ def putorderdetails(orderNumber):
 @orderdetail_routes.route('/orderdetails/<int:orderNumber>', methods=['DELETE'])
 def delorderdetails(orderNumber):
     found = orderNumber
+    row = session.query(Orderdetail).get(found)
+    if not row:
+        return resp.response_with(resp.SERVER_ERROR_404), resp.SERVER_ERROR_404
     try:
-        row = session.query(Orderdetail).get(found)
         session.delete(row)
         session.commit()
-        return resp.response_with(resp.SERVER_ERROR_404), resp.SERVER_ERROR_404
+        return resp.response_with(resp.SUCCESS_204, value={'message': 'Deleted Row'}), resp.SUCCESS_204
     except SQLAlchemyError as error:
         session.rollback()
         return resp.response_with(resp.BAD_REQUEST_400), resp.BAD_REQUEST_400
